src/test/java/itc/hoseo/hellospring/service/MemberServiceTest.java
package itc.hoseo.hellospring.service;

import itc.hoseo.hellospring.domain.Member;
import itc.hoseo.hellospring.repository.MemberRepository;
import lombok.RequiredArgsConstructor;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertThrows;

@SpringBootTest
public class MemberServiceTest {
    @Autowired
    private MemberService memberService;

    @Autowired
    private MemberRepository memberRepository;

    @Test
    public void testJoin(){

        assertThrows(RuntimeException.class, () ->{
            memberService.join(new Member("hello","hello","passwd"));
        });

        assertEquals(3, memberRepository.findAll().size());
    }

}